Updated SBT
Now on SBT 0.13.7-RC3. SBT 0.13.7 is broken, don't upgrade to
it.

SBT now caches lookups of dependencies. So much faster.

Add sbt-idea plugin as it generates better idea projects
(it links to sources and javadocs
 much better than idea just opening the sbt file)

Removed s-coverage plugin. We weren't using it and
intellij's built in coverage is probably better.